+++ Docs/CHANGELOG.md
@@ -48,6 +48,8 @@
     - Added support for MiscObjects (besides vehicles and pedestrians)
     - Reworked traffic signal handling: The name has to start now either with "id=" or "pos=" depending on whether the position or id is used as unique identifier
 ### :bug: Bug Fixes
+* Fixed bug causing RunningStopTest atomic criteria to trigger when lane changing near a STOP signal
+* Fixed bug causing RunningRedLightTest atomic criteria to occasionally not trigger
 * Fixed bug causing occasional frame_errors
 * Fixed #426: Avoid underground vehicles fall forever by disabling physics when spawning underground.
 * Fixed #427: Removed unnecessary warnings when using get_next_traffic_light() with non-cached locations